+++ index.js
@@ -1,26 +1,23 @@
 #!/usr/bin/env node
 
-const { inspect } = require('util')
-const config = require('./src/files')
-const analyse = require('./src/analyse')
+const files = require('./src/pipeline/files')
+const analyse = require('./src/pipeline/analyse')
+
 const cli = require('./src/reporters/cli')
+const github = require('./src/reporters/github')
+const build = require('./src/reporters/build')
+
+// old reporter
 const reporter = require('./src/reporter')
-const build = require('./src/build')
-
-const results = analyse(config)
-
 // reporter(report.files)
 
-cli.report(results)
-
-// exit with error code 1 if there are any failed checks
-if (results.status === 'fail') process.exit(1)
-
-process.on('unhandledRejection', function(reason) {
-  console.log('Unhandled Promise')
-  console.log(inspect(reason))
-  build.error()
-})
+try {
+  const results = analyse(files)
+  cli.report(results)
+  build.report(results)
+} catch (err) {
+  build.error(err)
+}
 
 /*
   This is the ideal structure to get to:
