+++ src/pipeline/analyse.js
